apps/api/requirements.txt
fastapi
uvicorn
sqlmodel
phonenumbers
apscheduler
vapi_server_sdk
python-dotenv
